Добавлена проверка на дубликат стихов

+++ src/main/kotlin/bot_chains/moderator/add_verse/SubmitAddVerse.kt
@@ -2,10 +2,7 @@ package bot_chains.moderator.add_verse
 
 import chain.Chain
 import core.Updating
-import data.ClearPoemModel
-import data.Poem
-import data.PoemStorage
-import data.PoemValidate
+import data.*
 import executables.AnswerToCallback
 import executables.DeleteMessage
 import executables.Executable
@@ -17,23 +14,30 @@ class SubmitAddVerse : Chain(OnCallbackGotten("submitAddVerse")) {
     override suspend fun executableChain(updating: Updating): List<Executable> {
         val poem = Poem(mStates.state(updating))
         return  if (poem.map(PoemValidate())) {
-            PoemStorage.Base.Instance().insertPoem(
-                poem
-            )
-            ClearPoemModel.Base(mStates, updating)
-            listOf(
-                AnswerToCallback(mKey, "Стих добавлен!"),
-                DeleteMessage(mKey, updating),
-                ModeratorMenu.Base(
-                    mKey,
-                    updating,
-                    false,
-                ) {
-                    mStates.state(updating).editor(mStates).apply {
-                        putInt("mainMessageId", it)
-                    }.commit()
-                }.message()
-            )
+            val poemIsExist = poem.map(DoesPoemExists(mKey))
+            if (poemIsExist.first) {
+                listOf(
+                    poemIsExist.second
+                )
+            } else {
+                PoemStorage.Base.Instance().insertPoem(
+                    poem
+                )
+                ClearPoemModel.Base(mStates, updating)
+                listOf(
+                    AnswerToCallback(mKey, "Стих добавлен!"),
+                    DeleteMessage(mKey, updating),
+                    ModeratorMenu.Base(
+                        mKey,
+                        updating,
+                        false,
+                    ) {
+                        mStates.state(updating).editor(mStates).apply {
+                            putInt("mainMessageId", it)
+                        }.commit()
+                    }.message()
+                )
+            }
         } else {
             listOf(
                 AnswerToCallback(mKey, "Заполните все поля!", true)
